requirements/common.txt
# Core Stuff
# -------------------------------------
Django==1.11.2
django-flat-responsive==1.4.1

# WSGI HTTP Server
# -------------------------------------
uWSGI==2.0.15

# Configuration
# -------------------------------------
argon2-cffi==16.3.0
django-environ==0.4.3
django-sites==0.9
python-dotenv==0.6.4

# Staticfiles
# -------------------------------------
whitenoise==3.3.0

# Extensions
# -------------------------------------
pytz==2017.2

# Models
# -------------------------------------
psycopg2==2.7.1

Pillow==4.1.1
django-versatileimagefield==1.7.1
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.6.3
django-rest-swagger==2.1.2

# Static files
# -------------------------------------
django-compressor==2.1.1
django-libsass==0.7
django-compressor-autoprefixer==0.1.0

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Raven is the Sentry client
# --------------------------
raven==6.1.0

# Third-party stuff goes here
-----------------------------
django-vote==2.1.4
